Create new AuthContext - manage token - manage user info

+++ src/helper/auth.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { matchPath } from 'react-router-dom';
 import { AUTH_API } from 'config';
 
-export const logout = async (authOnlyPaths, history, pathname, setToken, setTokenExpiry, setIsLoggedIn) => {
+export const logout = async (authOnlyPaths, history, pathname, dispatch) => {
   const isAuthOnlyPath = authOnlyPaths.find(({ path, exact, strict }) => matchPath(pathname, {
     path,
     exact,
@@ -12,17 +12,21 @@ export const logout = async (authOnlyPaths, history, pathname, setToken, setToke
   // remove refresh token cookie
   await axios.post(AUTH_API + '/logout', { withCredentials: true });
 
-  // to support logging out from all windows
-  localStorage.setItem('logout', Date.now());
-  localStorage.clear();
+  // This is to support logging out from all windows
+  // if 'hasLoggedIn' has a value then this means they haven't
+  // cleared the localStorage and triggered all logouts yet
+  // Conditional check to prevent bouncing logouts off tabs
+  if (localStorage.getItem('hasLoggedIn')) {
+    localStorage.setItem('logout', Date.now());
+    localStorage.clear();
+  }
 
   if (isAuthOnlyPath) {
     history.replace('/login');
   }
 
-  setToken(null);
-  setTokenExpiry(null);
-  setIsLoggedIn(false); // must be last after clearing tokens and everything
+  dispatch(['logout']);
+  console.log('%cLogged out!', 'color:purple');
 }
 
 export const getTokenFromRefresh = async () => {
@@ -33,23 +37,16 @@ export const getTokenFromRefresh = async () => {
         ok: true,
         token: data.token,
         tokenExpiry: data.tokenExpiry,
-        barista: {
-          id: data.id,
-          email: data.email,
-          displayName: data.displayName,
-          avatar: data.avatar,
-        },
       }
     }
     return { ok: false };
   } catch (err) {
-    console.log('ERROR GETTING REFRESH TOKEN:', err);
     return { ok: false };
   }
 }
 
 /**
- * authExchange options that doesn't require knowledge of userContext
+ * authExchange options that doesn't require knowledge of AuthContext
  */
 export const addAuthToOperation = ({
   authState,
@@ -81,6 +78,8 @@ export const didAuthError = ({ error }) => {
   const hasError = error.graphQLErrors.some(e => e.extensions?.code === 'invalid-jwt');
   return hasError;
 }
+
+export const willAuthError = ({ authState }) => !authState || new Date().getTime() >= new Date(authState.tokenExpiry).getTime();
 
 // {
 //   "errors": [
